write unit - ui test follow example

+++ app/src/main/java/com/example/mvvmApp/data/datasource/local/LocalDataSource.kt
@@ -10,6 +10,6 @@ interface LocalDataSource {
 
     fun getAllTableApps(): List<Table1App>
     fun getTableById(tableId: String): Table1App?
-    fun insertTableAppIfNotExists(table: Table1App): Int
+    fun insertTableAppIfNotExists(table: Table1App): Long
     fun deleteTableApp(table: Table1App): Int
 }
